[BugFix] Fixed authentication user flow bug.
Fixed bug where user had to login twice to get to app page.

The issue was that first login wasn't taken into account, due to page
manager not letting the time to spotify auth page to retrieve token info.
To fix this, a sleep function has been added to pageManager when current
route is spotify authentication.